[Refactor] Move primary key encoder to storage primitive layer
