+++ be/src/storage/lake/delta_writer.cpp
@@ -53,7 +53,7 @@
 #include "storage/load_spill_pipeline_merge_context.h"
 #include "storage/memtable.h"
 #include "storage/memtable_sink.h"
-#include "storage/primary_key_encoder.h"
+#include "storage/primitive/primary_key_encoder.h"
 #include "storage/storage_engine.h"
 #include "storage/storage_metrics.h"
