+++ be/src/storage/lake/lake_primary_index.cpp
@@ -28,7 +28,7 @@
 #include "storage/lake/rowset_update_state.h"
 #include "storage/lake/tablet.h"
 #include "storage/persistent_index_parallel_publish_context.h"
-#include "storage/primary_key_encoder.h"
+#include "storage/primitive/primary_key_encoder.h"
 #include "storage/tablet_meta_manager.h"
 
 namespace starrocks::lake {
