+++ be/src/storage/lake/lake_primary_key_recover.cpp
@@ -22,7 +22,7 @@
 #include "storage/lake/meta_file.h"
 #include "storage/lake/tablet.h"
 #include "storage/lake/update_manager.h"
-#include "storage/primary_key_encoder.h"
+#include "storage/primitive/primary_key_encoder.h"
 #include "storage/storage_engine.h"
 #include "storage/tablet_meta_manager.h"
