+++ be/src/storage/lake/pk_tablet_sst_writer.cpp
@@ -27,7 +27,7 @@
 #include "storage/lake/filenames.h"
 #include "storage/lake/tablet_manager.h"
 #include "storage/primary_index.h"
-#include "storage/primary_key_encoder.h"
+#include "storage/primitive/primary_key_encoder.h"
 #include "storage/rowset/segment_writer.h"
 #include "storage/sstable/table_builder.h"
