+++ be/src/storage/local_tablet_reader.cpp
@@ -20,7 +20,7 @@
 #include "serde/protobuf_serde.h"
 #include "storage/chunk_helper.h"
 #include "storage/primary_index.h"
-#include "storage/primary_key_encoder.h"
+#include "storage/primitive/primary_key_encoder.h"
 #include "storage/primitive/projection_iterator.h"
 #include "storage/storage_engine.h"
 #include "storage/tablet_manager.h"
