+++ be/src/storage/persistent_index.cpp
@@ -58,8 +58,8 @@
 #include "storage/persistent_index_parallel_publish_context.h"
 #include "storage/persistent_index_tablet_loader.h"
 #include "storage/primary_key_dump.h"
-#include "storage/primary_key_encoder.h"
 #include "storage/primitive/chunk_iterator.h"
+#include "storage/primitive/primary_key_encoder.h"
 #include "storage/rowset/rowset.h"
 #include "storage/storage_engine.h"
 #include "storage/tablet.h"
