+++ be/src/storage/primary_key_dump.cpp
@@ -25,8 +25,8 @@
 #include "storage/chunk_helper.h"
 #include "storage/del_vector.h"
 #include "storage/delta_column_group.h"
-#include "storage/primary_key_encoder.h"
 #include "storage/primitive/chunk_iterator.h"
+#include "storage/primitive/primary_key_encoder.h"
 #include "storage/rowset/rowset.h"
 #include "storage/rowset/segment_iterator.h"
 #include "storage/rowset/segment_options.h"
