+++ be/src/storage/primary_key_recover.cpp
@@ -17,7 +17,7 @@
 #include "column/chunk_factory.h"
 #include "serde/column_array_serde.h"
 #include "storage/chunk_helper.h"
-#include "storage/primary_key_encoder.h"
+#include "storage/primitive/primary_key_encoder.h"
 
 namespace starrocks {
