+++ be/src/storage/rowset_merger.cpp
@@ -28,8 +28,8 @@
 #include "gutil/stl_util.h"
 #include "storage/base/merge_iterator.h"
 #include "storage/chunk_helper.h"
-#include "storage/primary_key_encoder.h"
 #include "storage/primitive/empty_iterator.h"
+#include "storage/primitive/primary_key_encoder.h"
 #include "storage/primitive/union_iterator.h"
 #include "storage/rowset/column_reader.h"
 #include "storage/rowset/rowset_options.h"
