+++ be/src/storage/tablet_reader.cpp
@@ -28,7 +28,6 @@
 #include "common/system/backend_options.h"
 #include "gen_cpp/tablet_schema.pb.h"
 #include "gutil/stl_util.h"
-#include "primary_key_encoder.h"
 #include "runtime/type_info_allocator_adapter.h"
 #include "storage/aggregate_iterator.h"
 #include "storage/base/merge_iterator.h"
@@ -40,6 +39,7 @@
 #include "storage/primitive/column_predicate_factory.h"
 #include "storage/primitive/conjunctive_predicates.h"
 #include "storage/primitive/empty_iterator.h"
+#include "storage/primitive/primary_key_encoder.h"
 #include "storage/primitive/union_iterator.h"
 #include "storage/rowset/column_reader.h"
 #include "storage/rowset/rowid_range_option.h"
