+++ be/test/storage/rowset_merger_test.cpp
@@ -24,8 +24,8 @@
 #include "gutil/strings/substitute.h"
 #include "gutil/walltime.h"
 #include "storage/chunk_helper.h"
-#include "storage/primary_key_encoder.h"
 #include "storage/primitive/empty_iterator.h"
+#include "storage/primitive/primary_key_encoder.h"
 #include "storage/primitive/union_iterator.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_meta.h"
