+++ be/test/storage/tablet_updates_test.h
@@ -33,8 +33,8 @@
 #include "gutil/walltime.h"
 #include "storage/chunk_helper.h"
 #include "storage/kv_store.h"
-#include "storage/primary_key_encoder.h"
 #include "storage/primitive/empty_iterator.h"
+#include "storage/primitive/primary_key_encoder.h"
 #include "storage/primitive/union_iterator.h"
 #include "storage/rowset/rowset_factory.h"
 #include "storage/rowset/rowset_options.h"
